governance-app/requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
httpx==0.27.0
pyyaml==6.0.1
cryptography==42.0.5
python-dotenv==1.0.1
pydantic==2.7.1
